% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euctr_details.R
\name{euctr_details}
\alias{euctr_details}
\title{euctr_details}
\usage{
euctr_details(trn, n_retry = 10)
}
\arguments{
\item{trn}{A character vector containing a EUCTR trial registry
number}

\item{n_retry}{Integer number of times to try downloading again
before failing, default 10}
}
\value{
A list containing: 1. a logical TRUE/FALSE indicating
whether results are posted, 2. the trial start date, 3. a date
indicating the first version publication date and 4. the
"Global completion date" or the "Global end of trial date" if
the "Global completion date" is not reported, and NA if neither
is reported
}
\description{
Retrieves the start date, whether results were posted,
the publication date and the completion date for a EUCTR entry
}
\examples{
euctr_details("2012-001661-32") ## results
euctr_details("2020-005087-66") ## no results

}
