% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euctr_reg_dates.R
\name{euctr_reg_dates}
\alias{euctr_reg_dates}
\title{euctr_reg_dates}
\usage{
euctr_reg_dates(trn)
}
\arguments{
\item{trn}{A character vector containing a EUCTR trial registry
number}
}
\value{
A vector of dates
}
\description{
Retrieves the registration dates for a EUCTR entry
}
