% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euctr_results_posted.R
\name{euctr_results_posted}
\alias{euctr_results_posted}
\title{euctr_results_posted}
\usage{
euctr_results_posted(trn, n_retry = 10)
}
\arguments{
\item{trn}{A character vector containing a EUCTR trial registry
number}

\item{n_retry}{Integer number of times to try downloading again
before failing}
}
\value{
A list containing: 1. a logical TRUE/FALSE indicating
whether results are posted, 2. a date indicating the first
version publication date and 3. the "Global completion date" or
the "Global end of trial date" if the "Global completion date"
is not reported, and NA if neither is reported
}
\description{
Retrieves the date that results were first posted if
there is a link to trial results for a EUCTR entry
}
\examples{
euctr_results_posted("2012-001661-32") ## results
euctr_results_posted("2020-005087-66") ## no results

}
