% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euctr_start_date.R
\name{euctr_start_date}
\alias{euctr_start_date}
\title{euctr_start_date}
\usage{
euctr_start_date(trn, n_retry = 10)
}
\arguments{
\item{trn}{A character vector containing a EUCTR trial registry
number}

\item{n_retry}{A maximum number of times that the function should
try to download the page}
}
\value{
A character vector containing an ISO-8601 formatted date
}
\description{
Retrieves the start date for a EUCTR entry
}
